* fhandler_netdrive.cc: Use UINT32_MAX rather than UINT_MAX to make 	absolutely clear that we mean the 32 bit maximum. 	* fhandler_socket.cc: Ditto. 	* pinfo.cc: Ditto with INT32_MAX. 	* select.cc (fd_mask): Drop gratuitious definition. 	(sizeof_fd_set): Cast to size_t. 	* shared.cc (pround): Cast to ptrdiff_t. 	* strace.cc (strace::activate): Fix format string. 	(strace::write_childpid): Take pid_t as parameter. 	* include/sys/strace.h (strace::write_childpid): Change declaration 	accordingly. 	* sync.cc (muto::init): Cast in call to InterlockedExchangePointer to 	make compiler happy. 	* sync.h (class muto): De-const name. 	* syslog.cc (try_connect_syslogd): Take size_t as third argument. 	(vsyslog): Convert len to size_t. 	* thread.cc: Use {U}INT32_MAX rather than {U}LONG_MAX.  Througout change 	types used in Windows calls to matching Windows types. 	(pthread::cancel): Fix access to instruction pointer for x86_64. 	(pthread_attr_getstack): Cast to ptrdiff_t for pointer arithmetic. 	* thread.h: Throughout convert 'long' types to uint32_t or LONG.

+++ winsup/cygwin/syslog.cc
@@ -249,7 +249,7 @@ connect_syslogd ()
 }
 
 static int
-try_connect_syslogd (int priority, const char *msg, int len)
+try_connect_syslogd (int priority, const char *msg, size_t len)
 {
   ssize_t ret = -1;
 
@@ -402,7 +402,7 @@ vsyslog (int priority, const char *message, va_list ap)
 
     }
   char *total_msg = pass.get_message ();
-  int len = strlen (total_msg);
+  size_t len = strlen (total_msg);
   if (len != 0 && (total_msg[len - 1] == '\n'))
     total_msg[--len] = '\0';
